+++ test/src/hosts/host_test.dart
@@ -1,0 +1,34 @@
+// import 'package:test/test.dart';
+// import 'package:uhst/src/contracts/uhst_api_client.dart';
+// import 'package:uhst/src/contracts/uhst_host_event.dart';
+// import 'package:uhst/src/contracts/uhst_socket_provider.dart';
+// import 'package:uhst/uhst.dart';
+
+// var host = UhstHost.create(
+//     apiClient: {} as UhstApiClient,
+//     socketProvider: {} as UhstSocketProvider,
+//     debug: false);
+
+// /// FIXME: This test is not working:(
+// /// somehow there is an error with loading uhst_host
+// /// It is not connected with universal_html
+// /// (after reomval there is still no result)
+// /// and need to be investigated
+// void main() {
+//   group('# host', () {
+//     test('init host', () {
+//       // expect(host.h.eventStream, isNotNull);
+//       // expect(host.h.eventStreamController, isNotNull);
+//       // expect(host.h.debug, equals(false));
+//       // expect(host.h.apiClient, isNotNull);
+//     });
+//     test('send and catch message inside event stream', () {
+//       // var message = 'testMessage';
+//       var messageType = HostEventType.ready;
+//       // host.h.emit(message: HostEventType.ready, body: message);
+//       // host.h.eventStream.listen((event) {
+//       //   expect(event.entries, equals({messageType: message}));
+//       // });
+//     });
+//   });
+// }