+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - fix: The package description is too short.
 - fix: The value of the local variable 'onOpenSubcription' isn't used (api_client).
+- fix: cancel subscribe warning.
 
 # 0.0.1-nullsafety.0