+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - fix: The package description is too short.
 - fix: The value of the local variable 'onOpenSubcription' isn't used (api_client).
 - fix: cancel subscribe warning.
+- feat: test for jwt util
 
 # 0.0.1-nullsafety.0