+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.0.1-nullsafety.1
+
+- fix: The package description is too short.
+- fix: The value of the local variable 'onOpenSubcription' isn't used (api_client).
+- fix: cancel subscribe warning.
+- feat: test for jwt util
+- feat: type definitions test
+- fix: payload type conversion
+- fix: typing for uhst host create
+
 # 0.0.1-nullsafety.0
 
 - Initial development release.